refactored App into Controller and added We files
