+++ apps/applet/ui/package.json
@@ -1,0 +1,80 @@
+{
+  "name": "we-applet",
+  "description": "Webcomponent lit-element following open-wc recommendations",
+  "license": "MIT",
+  "author": "lit-element",
+  "version": "0.0.0",
+  "scripts": {
+    "lint": "eslint --ext .ts,.html . --ignore-path .gitignore && prettier \"**/*.ts\" --check --ignore-path .gitignore",
+    "format": "eslint --ext .ts,.html . --fix --ignore-path .gitignore && prettier \"**/*.ts\" --write --ignore-path .gitignore",
+    "build": "rimraf dist && tsc && rollup -c rollup.config.js",
+    "package": "npm run build:happ && npm run build && cd dist && bestzip ../dist.zip * && cd .. && hc web-app pack ../workdir",
+    "demo": "npm run build:happ && npm run build && cross-env HC_PORT=$(port) ADMIN_PORT=$(port) concurrently \"npm run start:happ\" \"wds --config web-dev-server.config.mjs\"",
+    "start:happ": "RUST_LOG=warn echo \"pass\" | hc s --piped -f=$ADMIN_PORT generate ../workdir/talking-stickies.happ --run=$HC_PORT -a talking-stickies network mdns",
+    "build:happ": "npm run build:dnas && hc app pack ../workdir",
+    "build:dnas": "npm run build:zomes && hc dna pack ../dnas/talking-stickies",
+    "build:zomes": "CARGO_TARGET_DIR=../../../target cargo build --release --target wasm32-unknown-unknown"
+  },
+  "dependencies": {
+    "@holochain-open-dev/cell-client": "^0.7.2",
+    "@holochain-open-dev/core-types": "^0.4.1",
+    "@holochain/client": "^0.7.0",
+    "@holo-host/talking-stickies": "^0.3.0",
+    "@lightningrodlabs/we-applet": "^0.1.0",
+    "@lit-labs/context": "^0.1.2",
+    "@open-wc/scoped-elements": "^2.0.1",
+    "@scoped-elements/material-web": "^0.0.19",
+    "lit": "^2.0.0",
+    "lit-svelte-stores": "^0.2.0",
+    "svelte": "^3.44.1"
+  },
+  "devDependencies": {
+    "@babel/preset-env": "^7.15.0",
+    "@open-wc/building-rollup": "^1.10.0",
+    "@open-wc/eslint-config": "^4.3.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "18.0.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^13.0.4",
+    "@rollup/plugin-replace": "^3.0.0",
+    "@rollup/plugin-typescript": "^8.3.2",
+    "@typescript-eslint/eslint-plugin": "^4.29.2",
+    "@typescript-eslint/parser": "^4.29.2",
+    "@web/dev-server": "^0.1.21",
+    "@web/dev-server-rollup": "^0.3.10",
+    "@web/rollup-plugin-html": "^1.9.1",
+    "@web/rollup-plugin-import-meta-assets": "^1.0.7",
+    "babel-plugin-template-html-minifier": "^4.1.0",
+    "concurrently": "^5.3.0",
+    "deepmerge": "^4.2.2",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "husky": "^4.3.8",
+    "lint-staged": "^10.5.4",
+    "prettier": "^2.3.2",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.56.2",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-workbox": "^6.2.0",
+    "run-singleton-cli": "^0.0.5",
+    "tslib": "^2.3.1",
+    "typescript": "^4.3.5"
+  },
+  "prettier": {
+    "singleQuote": true,
+    "arrowParens": "avoid"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.ts": [
+      "eslint --fix",
+      "prettier --write"
+    ]
+  }
+}
